Ajust de tipos das colunas

+++ models/raw/cgm/osinfo_rh/raw_osinfo_rh__dc_grau_instrucao.sql
@@ -7,9 +7,9 @@
 }}
 
 select
-    safe_cast(`GINS_CD_GRAU_INSTRUCAO` as integer) as grau_instrucao_codigo,
+    safe_cast(`GINS_CD_GRAU_INSTRUCAO` as int64) as grau_instrucao_codigo,
     safe_cast(`GINS_DS_GRAU_INSTRUCAO` as string) as grau_instrucao_descricao,
-    safe_cast(`GINS_ORDEM` as integer) as grau_instrucao_ordem,
+    safe_cast(`GINS_ORDEM` as int64) as grau_instrucao_ordem,
     DATETIME(_prefect_extracted_at, "America/Sao_Paulo") AS datalake_loaded_at,
     DATETIME(current_timestamp(), "America/Sao_Paulo") AS datalake_transformed_at
 FROM {{ source('brutos_osinfo_rh_staging', 'dc_grau_instrucao') }}
