+++ models/raw/cgm/osinfo_rh/raw_osinfo_rh__dc_raca_cor.sql
@@ -7,7 +7,7 @@
 }}
 
 select
-    safe_cast(`RACO_CD_RACA_COR` as integer) as raca_cor_codigo,
+    safe_cast(`RACO_CD_RACA_COR` as int64) as raca_cor_codigo,
     safe_cast(`RACO_DS_DESCRICAO` as string) as raca_cor_descricao,
     DATETIME(_prefect_extracted_at, "America/Sao_Paulo") AS datalake_loaded_at,
     DATETIME(current_timestamp(), "America/Sao_Paulo") AS datalake_transformed_at
