+++ models/raw/cgm/osinfo_rh/raw_osinfo_rh__dc_sexo.sql
@@ -7,7 +7,7 @@
 }}
 
 select
-    safe_cast(`SEX_CD_SEXO` as integer) as sexo_codigo,
+    safe_cast(`SEX_CD_SEXO` as int64) as sexo_codigo,
     safe_cast(`SEX_DS_DESCRICAO` as string) as sexo_descricao,
     DATETIME(_prefect_extracted_at, "America/Sao_Paulo") AS datalake_loaded_at,
     DATETIME(current_timestamp(), "America/Sao_Paulo") AS datalake_transformed_at
