+++ models/raw/cgm/osinfo_rh/raw_osinfo_rh__tb_cbo.sql
@@ -9,7 +9,7 @@
 select
     safe_cast(`CBO_CD_CBO` as string) as cbo_codigo,
     safe_cast(`CBO_DS_CBO` as string) as cbo_descricao,
-    safe_cast(`CBO_IN_ATIVIDADE_FIM` as integer) as cbo_atividade_fim,
+    safe_cast(`CBO_IN_ATIVIDADE_FIM` as int64) as cbo_atividade_fim,
     DATETIME(_prefect_extracted_at, "America/Sao_Paulo") AS datalake_loaded_at,
     DATETIME(current_timestamp(), "America/Sao_Paulo") AS datalake_transformed_at
 FROM {{ source('brutos_osinfo_rh_staging', 'tb_cbo') }}
