+++ models/raw/cgm/osinfo_rh/raw_osinfo_rh__tb_grupo.sql
@@ -9,7 +9,7 @@ select
     safe_cast(`RHGR_TP_TIPO` as string) as grupo_tipo,
     safe_cast(`RHGR_CD_CODIGO` as string) as grupo_codigo,
     safe_cast(`RHGR_NM_NOME` as string) as grupo_nome,
-    safe_cast(`RHGR_IN_TOTALIZAR` as int) as grupo_totalizar,
+    safe_cast(`RHGR_IN_TOTALIZAR` as int64) as grupo_totalizar,
     DATETIME(_prefect_extracted_at, "America/Sao_Paulo") AS datalake_loaded_at,
     DATETIME(current_timestamp(), "America/Sao_Paulo") AS datalake_transformed_at
 FROM {{ source('brutos_osinfo_rh_staging', 'tb_grupo') }}
